refactor(module:date-picker): merge picker and date-picker component
