+++ components/date-picker/date-picker.component.ts
@@ -4,6 +4,7 @@
  */
 
 import {
+  AfterViewInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
@@ -11,34 +12,51 @@ import {
   EventEmitter,
   forwardRef,
   Host,
+  Inject,
   Input,
   OnChanges,
   OnDestroy,
   OnInit,
   Optional,
   Output,
+  QueryList,
   Renderer2,
   SimpleChanges,
   TemplateRef,
   ViewChild,
+  ViewChildren,
   ViewEncapsulation
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
+import { slideMotion } from 'ng-zorro-antd/core/animation';
 import { NzNoAnimationDirective } from 'ng-zorro-antd/core/no-animation';
-import { CandyDate, cloneDate, CompatibleValue } from 'ng-zorro-antd/core/time';
+import { NzResizeObserver } from 'ng-zorro-antd/core/resize-observers';
+import { CandyDate, cloneDate, CompatibleValue, wrongSortOrder } from 'ng-zorro-antd/core/time';
 import { BooleanInput, FunctionProp, NzSafeAny, OnChangeType, OnTouchedType } from 'ng-zorro-antd/core/types';
 import { InputBoolean, toBoolean, valueFunctionProp } from 'ng-zorro-antd/core/util';
 import { DateHelperService, NzDatePickerI18nInterface, NzDatePickerLangI18nInterface, NzI18nService } from 'ng-zorro-antd/i18n';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { DatePickerService } from './date-picker.service';
+import { DateRangePopupComponent } from './date-range-popup.component';
+import { PREFIX_CLASS } from './util';
 
 import { Direction, Directionality } from '@angular/cdk/bidi';
+import { ESCAPE } from '@angular/cdk/keycodes';
+import {
+  CdkConnectedOverlay,
+  CdkOverlayOrigin,
+  ConnectedOverlayPositionChange,
+  ConnectionPositionPair,
+  HorizontalConnectionPos,
+  VerticalConnectionPos
+} from '@angular/cdk/overlay';
+import { Platform } from '@angular/cdk/platform';
+import { DOCUMENT } from '@angular/common';
 import { NzConfigKey, NzConfigService, WithConfig } from 'ng-zorro-antd/core/config';
-import { NzPickerComponent } from './picker.component';
-import { CompatibleDate, DisabledTimeFn, NzDateMode, PresetRanges, SupportTimeOptions } from './standard-types';
+import { CompatibleDate, DisabledTimeFn, NzDateMode, PresetRanges, RangePartType, SupportTimeOptions } from './standard-types';
 
-const POPUP_STYLE_PATCH = { position: 'relative' }; // Aim to override antd's style to support overlay's position strategy (position:absolute will cause it not working beacuse the overlay can't get the height/width of it's content)
+const POPUP_STYLE_PATCH = { position: 'relative' }; // Aim to override antd's style to support overlay's position strategy (position:absolute will cause it not working because the overlay can't get the height/width of it's content)
 const NZ_CONFIG_MODULE_NAME: NzConfigKey = 'datePicker';
 
 export type NzDatePickerSizeType = 'large' | 'default' | 'small';
@@ -52,54 +70,131 @@ export type NzDatePickerSizeType = 'large' | 'default' | 'small';
   selector: 'nz-date-picker,nz-week-picker,nz-month-picker,nz-year-picker,nz-range-picker',
   exportAs: 'nzDatePicker',
   template: `
-    <div
-      nz-picker
-      style="display: inherit; align-items: center; width: 100%;"
-      [id]="nzId"
-      [isRange]="isRange"
-      [open]="nzOpen"
-      [dir]="dir"
-      [separator]="nzSeparator"
-      [disabled]="nzDisabled"
-      [inputReadOnly]="nzInputReadOnly"
-      [inline]="nzInline"
-      [format]="nzFormat"
-      [allowClear]="nzAllowClear"
-      [autoFocus]="nzAutoFocus"
-      [placeholder]="nzPlaceHolder"
-      [dropdownClassName]="nzDropdownClassName"
-      [class.ant-picker-dropdown-rtl]="dir === 'rtl'"
-      [popupStyle]="nzPopupStyle"
-      [noAnimation]="!!noAnimation?.nzNoAnimation"
-      [suffixIcon]="nzSuffixIcon"
-      [hasBackdrop]="nzBackdrop"
-      (openChange)="onOpenChange($event)"
-      (focusChange)="onFocusChange($event)"
-      [nzId]="nzId"
+    <ng-container *ngIf="!nzInline; else inlineMode">
+      <!-- Content of single picker -->
+      <div *ngIf="!isRange" class="{{ prefixCls }}-input">
+        <input
+          #pickerInput
+          [attr.id]="nzId"
+          [class.ant-input-disabled]="nzDisabled"
+          [disabled]="nzDisabled"
+          [readOnly]="nzInputReadOnly"
+          [(ngModel)]="inputValue"
+          placeholder="{{ getPlaceholder() }}"
+          [size]="inputSize"
+          (focus)="onFocus($event)"
+          (focusout)="onFocusout($event)"
+          (ngModelChange)="onInputChange($event)"
+          (keyup.enter)="onKeyupEnter($event)"
+        />
+        <ng-container *ngTemplateOutlet="tplRightRest"></ng-container>
+      </div>
+
+      <!-- Content of range picker -->
+      <ng-container *ngIf="isRange">
+        <div class="{{ prefixCls }}-input">
+          <ng-container *ngTemplateOutlet="tplRangeInput; context: { partType: 'left' }"></ng-container>
+        </div>
+        <div #separatorElement class="{{ prefixCls }}-range-separator">
+          <span class="{{ prefixCls }}-separator">
+            <ng-container *ngIf="nzSeparator; else defaultSeparator">{{ nzSeparator }}</ng-container>
+          </span>
+          <ng-template #defaultSeparator>
+            <i nz-icon nzType="swap-right" nzTheme="outline"></i>
+          </ng-template>
+        </div>
+        <div class="{{ prefixCls }}-input">
+          <ng-container *ngTemplateOutlet="tplRangeInput; context: { partType: 'right' }"></ng-container>
+        </div>
+        <ng-container *ngTemplateOutlet="tplRightRest"></ng-container>
+      </ng-container>
+    </ng-container>
+    <!-- Input for Range ONLY -->
+    <ng-template #tplRangeInput let-partType="partType">
+      <input
+        #rangePickerInput
+        [disabled]="nzDisabled"
+        [readOnly]="nzInputReadOnly"
+        [size]="inputSize"
+        (click)="onClickInputBox($event)"
+        (focusout)="onFocusout($event)"
+        (focus)="onFocus($event, partType)"
+        (keyup.enter)="onKeyupEnter($event)"
+        [(ngModel)]="inputValue[datePickerService.getActiveIndex(partType)]"
+        (ngModelChange)="onInputChange($event)"
+        placeholder="{{ getPlaceholder(partType) }}"
+      />
+    </ng-template>
+
+    <!-- Right operator icons -->
+    <ng-template #tplRightRest>
+      <div class="{{ prefixCls }}-active-bar" [ngStyle]="activeBarStyle"></div>
+      <span *ngIf="showClear()" class="{{ prefixCls }}-clear" (click)="onClickClear($event)">
+        <i nz-icon nzType="close-circle" nzTheme="fill"></i>
+      </span>
+      <span class="{{ prefixCls }}-suffix">
+        <ng-container *nzStringTemplateOutlet="nzSuffixIcon; let suffixIcon">
+          <i nz-icon [nzType]="suffixIcon"></i>
+        </ng-container>
+      </span>
+    </ng-template>
+
+    <ng-template #inlineMode>
+      <div class="ant-picker-wrapper" [nzNoAnimation]="!!noAnimation?.nzNoAnimation" [@slideMotion]="'enter'" style="position: relative;">
+        <div
+          class="{{ prefixCls }}-dropdown {{ nzDropdownClassName }}"
+          [class.ant-picker-dropdown-rtl]="dir === 'rtl'"
+          [class.ant-picker-dropdown-placement-bottomLeft]="currentPositionY === 'bottom' && currentPositionX === 'start'"
+          [class.ant-picker-dropdown-placement-topLeft]="currentPositionY === 'top' && currentPositionX === 'start'"
+          [class.ant-picker-dropdown-placement-bottomRight]="currentPositionY === 'bottom' && currentPositionX === 'end'"
+          [class.ant-picker-dropdown-placement-topRight]="currentPositionY === 'top' && currentPositionX === 'end'"
+          [class.ant-picker-dropdown-range]="isRange"
+          [class.ant-picker-active-left]="datePickerService.activeInput === 'left'"
+          [class.ant-picker-active-right]="datePickerService.activeInput === 'right'"
+          [ngStyle]="nzPopupStyle"
+        >
+          <date-range-popup
+            [isRange]="isRange"
+            [inline]="nzInline"
+            [defaultPickerValue]="nzDefaultPickerValue"
+            [showWeek]="nzMode === 'week'"
+            [panelMode]="panelMode"
+            (panelModeChange)="onPanelModeChange($event)"
+            (calendarChange)="onCalendarChange($event)"
+            [locale]="nzLocale?.lang!"
+            [showToday]="nzMode === 'date' && nzShowToday && !isRange && !nzShowTime"
+            [showNow]="nzMode === 'date' && nzShowNow && !isRange && !!nzShowTime"
+            [showTime]="nzShowTime"
+            [dateRender]="nzDateRender"
+            [disabledDate]="nzDisabledDate"
+            [disabledTime]="nzDisabledTime"
+            [extraFooter]="extraFooter"
+            [ranges]="nzRanges"
+            [dir]="dir"
+            (resultOk)="onResultOk()"
+          ></date-range-popup>
+        </div>
+      </div>
+    </ng-template>
+
+    <!-- Overlay -->
+    <ng-template
+      cdkConnectedOverlay
+      nzConnectedOverlay
+      [cdkConnectedOverlayHasBackdrop]="nzBackdrop"
+      [cdkConnectedOverlayOrigin]="origin"
+      [cdkConnectedOverlayOpen]="realOpenState"
+      [cdkConnectedOverlayPositions]="overlayPositions"
+      [cdkConnectedOverlayTransformOriginOn]="'.ant-picker-wrapper'"
+      (positionChange)="onPositionChange($event)"
+      (detach)="close()"
+      (overlayKeydown)="onOverlayKeydown($event)"
     >
-      <date-range-popup
-        [isRange]="isRange"
-        [inline]="nzInline"
-        [defaultPickerValue]="nzDefaultPickerValue"
-        [showWeek]="nzMode === 'week'"
-        [panelMode]="panelMode"
-        (panelModeChange)="onPanelModeChange($event)"
-        (calendarChange)="onCalendarChange($event)"
-        [locale]="nzLocale?.lang!"
-        [showToday]="nzMode === 'date' && nzShowToday && !isRange && !nzShowTime"
-        [showNow]="nzMode === 'date' && nzShowNow && !isRange && !!nzShowTime"
-        [showTime]="nzShowTime"
-        [dateRender]="nzDateRender"
-        [disabledDate]="nzDisabledDate"
-        [disabledTime]="nzDisabledTime"
-        [extraFooter]="extraFooter"
-        [ranges]="nzRanges"
-        [dir]="dir"
-        (resultOk)="onResultOk()"
-      ></date-range-popup>
-    </div>
+      <ng-container *ngTemplateOutlet="inlineMode"></ng-container>
+    </ng-template>
   `,
   host: {
+    '[class.ant-picker]': `true`,
     '[class.ant-picker-range]': `isRange`,
     '[class.ant-picker-large]': `nzSize === 'large'`,
     '[class.ant-picker-small]': `nzSize === 'small'`,
@@ -107,7 +202,7 @@ export type NzDatePickerSizeType = 'large' | 'default' | 'small';
     '[class.ant-picker-rtl]': `dir === 'rtl'`,
     '[class.ant-picker-borderless]': `nzBorderless`,
     '[class.ant-picker-inline]': `nzInline`,
-    '(click)': 'picker.onClickInputBox($event)'
+    '(click)': 'onClickInputBox($event)'
   },
   providers: [
     DatePickerService,
@@ -116,9 +211,10 @@ export type NzDatePickerSizeType = 'large' | 'default' | 'small';
       multi: true,
       useExisting: forwardRef(() => NzDatePickerComponent)
     }
-  ]
+  ],
+  animations: [slideMotion]
 })
-export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, ControlValueAccessor {
+export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, AfterViewInit, ControlValueAccessor {
   readonly _nzModuleName: NzConfigKey = NZ_CONFIG_MODULE_NAME;
   static ngAcceptInputType_nzAllowClear: BooleanInput;
   static ngAcceptInputType_nzAutoFocus: BooleanInput;
@@ -176,8 +272,6 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
   @Output() readonly nzOnOk = new EventEmitter<CompatibleDate | null>();
   @Output() readonly nzOnOpenChange = new EventEmitter<boolean>();
 
-  @ViewChild(NzPickerComponent, { static: true }) picker!: NzPickerComponent;
-
   @Input() get nzShowTime(): SupportTimeOptions | boolean {
     return this.showTime;
   }
@@ -186,6 +280,287 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
     this.showTime = typeof value === 'object' ? value : toBoolean(value);
   }
 
+  // ------------------------------------------------------------------------
+  // Input API Start
+  // ------------------------------------------------------------------------
+  @ViewChild(CdkConnectedOverlay, { static: false }) cdkConnectedOverlay?: CdkConnectedOverlay;
+  @ViewChild(DateRangePopupComponent, { static: false }) panel!: DateRangePopupComponent;
+  @ViewChild('separatorElement', { static: false }) separatorElement?: ElementRef;
+  @ViewChild('pickerInput', { static: false }) pickerInput?: ElementRef<HTMLInputElement>;
+  @ViewChildren('rangePickerInput') rangePickerInputs?: QueryList<ElementRef<HTMLInputElement>>;
+
+  origin: CdkOverlayOrigin;
+  document: Document;
+  inputSize: number = 12;
+  inputWidth?: number;
+  destroy$ = new Subject();
+  prefixCls = PREFIX_CLASS;
+  inputValue!: NzSafeAny;
+  activeBarStyle: object = {};
+  overlayOpen: boolean = false; // Available when "nzOpen" = undefined
+  overlayPositions: ConnectionPositionPair[] = [
+    {
+      offsetY: 2,
+      originX: 'start',
+      originY: 'bottom',
+      overlayX: 'start',
+      overlayY: 'top'
+    },
+    {
+      offsetY: -2,
+      originX: 'start',
+      originY: 'top',
+      overlayX: 'start',
+      overlayY: 'bottom'
+    },
+    {
+      offsetY: 2,
+      originX: 'end',
+      originY: 'bottom',
+      overlayX: 'end',
+      overlayY: 'top'
+    },
+    {
+      offsetY: -2,
+      originX: 'end',
+      originY: 'top',
+      overlayX: 'end',
+      overlayY: 'bottom'
+    }
+  ] as ConnectionPositionPair[];
+  currentPositionX: HorizontalConnectionPos = 'start';
+  currentPositionY: VerticalConnectionPos = 'bottom';
+
+  get realOpenState(): boolean {
+    // The value that really decide the open state of overlay
+    return this.isOpenHandledByUser() ? !!this.nzOpen : this.overlayOpen;
+  }
+
+  ngAfterViewInit(): void {
+    if (this.nzAutoFocus) {
+      this.focus();
+    }
+
+    if (this.isRange && this.platform.isBrowser) {
+      this.nzResizeObserver
+        .observe(this.elementRef)
+        .pipe(takeUntil(this.destroy$))
+        .subscribe(() => {
+          this.updateInputWidthAndArrowLeft();
+        });
+    }
+
+    this.datePickerService.inputPartChange$.pipe(takeUntil(this.destroy$)).subscribe(partType => {
+      if (partType) {
+        this.datePickerService.activeInput = partType;
+      }
+      this.focus();
+      this.updateInputWidthAndArrowLeft();
+    });
+  }
+
+  updateInputWidthAndArrowLeft(): void {
+    this.inputWidth = this.rangePickerInputs?.first?.nativeElement.offsetWidth || 0;
+
+    const baseStyle = { position: 'absolute', width: `${this.inputWidth}px` };
+    this.datePickerService.arrowLeft =
+      this.datePickerService.activeInput === 'left' ? 0 : this.inputWidth + this.separatorElement?.nativeElement.offsetWidth || 0;
+
+    if (this.dir === 'rtl') {
+      this.activeBarStyle = { ...baseStyle, right: `${this.datePickerService.arrowLeft}px` };
+    } else {
+      this.activeBarStyle = { ...baseStyle, left: `${this.datePickerService.arrowLeft}px` };
+    }
+
+    this.cdr.markForCheck();
+  }
+
+  getInput(partType?: RangePartType): HTMLInputElement | undefined {
+    if (this.nzInline) {
+      return undefined;
+    }
+    return this.isRange
+      ? partType === 'left'
+        ? this.rangePickerInputs?.first.nativeElement
+        : this.rangePickerInputs?.last.nativeElement
+      : this.pickerInput!.nativeElement;
+  }
+
+  focus(): void {
+    const activeInputElement = this.getInput(this.datePickerService.activeInput);
+    if (this.document.activeElement !== activeInputElement) {
+      activeInputElement?.focus();
+    }
+  }
+
+  onFocus(event: FocusEvent, partType?: RangePartType): void {
+    event.preventDefault();
+    if (partType) {
+      this.datePickerService.inputPartChange$.next(partType);
+    }
+    this.renderClass(true);
+  }
+
+  // blur event has not the relatedTarget in IE11, use focusout instead.
+  onFocusout(event: FocusEvent): void {
+    event.preventDefault();
+    if (!this.elementRef.nativeElement.contains(event.relatedTarget)) {
+      this.checkAndClose();
+    }
+    this.renderClass(false);
+  }
+
+  // Show overlay content
+  open(): void {
+    if (this.nzInline) {
+      return;
+    }
+    if (!this.realOpenState && !this.nzDisabled) {
+      this.updateInputWidthAndArrowLeft();
+      this.overlayOpen = true;
+      this.nzOnOpenChange.emit(true);
+      this.cdr.markForCheck();
+    }
+  }
+
+  close(): void {
+    if (this.nzInline) {
+      return;
+    }
+    if (this.realOpenState) {
+      this.overlayOpen = false;
+      this.nzOnOpenChange.emit(false);
+    }
+  }
+
+  showClear(): boolean {
+    return !this.nzDisabled && !this.isEmptyValue(this.datePickerService.value) && this.nzAllowClear;
+  }
+
+  checkAndClose(): void {
+    if (!this.realOpenState) {
+      return;
+    }
+
+    if (this.panel.isAllowed(this.datePickerService.value!, true)) {
+      if (Array.isArray(this.datePickerService.value) && wrongSortOrder(this.datePickerService.value)) {
+        const index = this.datePickerService.getActiveIndex();
+        const value = this.datePickerService.value[index];
+        this.panel.changeValueFromSelect(value!, true);
+        return;
+      }
+      this.updateInputValue();
+      this.datePickerService.emitValue$.next();
+    } else {
+      this.datePickerService.setValue(this.datePickerService.initialValue!);
+      this.close();
+    }
+  }
+
+  onClickInputBox(event: MouseEvent): void {
+    event.stopPropagation();
+    this.focus();
+    if (!this.isOpenHandledByUser()) {
+      this.open();
+    }
+  }
+
+  onOverlayKeydown(event: KeyboardEvent): void {
+    if (event.keyCode === ESCAPE) {
+      this.datePickerService.initValue();
+    }
+  }
+
+  // NOTE: A issue here, the first time position change, the animation will not be triggered.
+  // Because the overlay's "positionChange" event is emitted after the content's full shown up.
+  // All other components like "nz-dropdown" which depends on overlay also has the same issue.
+  // See: https://github.com/NG-ZORRO/ng-zorro-antd/issues/1429
+  onPositionChange(position: ConnectedOverlayPositionChange): void {
+    this.currentPositionX = position.connectionPair.originX;
+    this.currentPositionY = position.connectionPair.originY;
+    this.cdr.detectChanges(); // Take side-effects to position styles
+  }
+
+  onClickClear(event: MouseEvent): void {
+    event.preventDefault();
+    event.stopPropagation();
+
+    this.datePickerService.initValue(true);
+    this.datePickerService.emitValue$.next();
+  }
+
+  updateInputValue(): void {
+    const newValue = this.datePickerService.value;
+    if (this.isRange) {
+      this.inputValue = newValue ? (newValue as CandyDate[]).map(v => this.formatValue(v)) : ['', ''];
+    } else {
+      this.inputValue = this.formatValue(newValue as CandyDate);
+    }
+    this.cdr.markForCheck();
+  }
+
+  formatValue(value: CandyDate): string {
+    return this.dateHelper.format(value && (value as CandyDate).nativeDate, this.nzFormat);
+  }
+
+  onInputChange(value: string, isEnter: boolean = false): void {
+    /**
+     * in IE11 focus/blur will trigger ngModelChange if placeholder changes,
+     * so we forbidden IE11 to open panel through input change
+     */
+    if (
+      !this.platform.TRIDENT &&
+      this.document.activeElement === this.getInput(this.datePickerService.activeInput) &&
+      !this.realOpenState
+    ) {
+      this.open();
+      return;
+    }
+
+    const date = this.checkValidDate(value);
+    // Can only change date when it's open
+    if (date && this.realOpenState) {
+      this.panel.changeValueFromSelect(date, isEnter);
+    }
+  }
+
+  onKeyupEnter(event: Event): void {
+    this.onInputChange((event.target as HTMLInputElement).value, true);
+  }
+
+  private checkValidDate(value: string): CandyDate | null {
+    const date = new CandyDate(this.dateHelper.parseDate(value, this.nzFormat));
+
+    if (!date.isValid() || value !== this.dateHelper.format(date.nativeDate, this.nzFormat)) {
+      return null;
+    }
+
+    return date;
+  }
+
+  getPlaceholder(partType?: RangePartType): string {
+    return this.isRange ? this.nzPlaceHolder[this.datePickerService.getActiveIndex(partType!)] : (this.nzPlaceHolder as string);
+  }
+
+  isEmptyValue(value: CompatibleValue): boolean {
+    if (value === null) {
+      return true;
+    } else if (this.isRange) {
+      return !value || !Array.isArray(value) || value.every(val => !val);
+    } else {
+      return !value;
+    }
+  }
+
+  // Whether open state is permanently controlled by user himself
+  isOpenHandledByUser(): boolean {
+    return this.nzOpen !== undefined;
+  }
+
+  // ------------------------------------------------------------------------
+  // Input API End
+  // ------------------------------------------------------------------------
+
   constructor(
     public nzConfigService: NzConfigService,
     public datePickerService: DatePickerService,
@@ -193,12 +568,15 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
     protected cdr: ChangeDetectorRef,
     private renderer: Renderer2,
     private elementRef: ElementRef,
-    protected dateHelper: DateHelperService,
+    private dateHelper: DateHelperService,
+    private nzResizeObserver: NzResizeObserver,
+    private platform: Platform,
+    @Inject(DOCUMENT) doc: NzSafeAny,
     @Optional() private directionality: Directionality,
     @Host() @Optional() public noAnimation?: NzNoAnimationDirective
   ) {
-    // TODO: move to host after View Engine deprecation
-    this.elementRef.nativeElement.classList.add('ant-picker');
+    this.document = doc;
+    this.origin = new CdkOverlayOrigin(this.elementRef);
   }
 
   ngOnInit(): void {
@@ -232,13 +610,17 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
       this.close();
     });
 
-    this.setModeAndFormat();
-
     this.directionality.change?.pipe(takeUntil(this.destroyed$)).subscribe((direction: Direction) => {
       this.dir = direction;
       this.cdr.detectChanges();
     });
     this.dir = this.directionality.value;
+    this.inputValue = this.isRange ? ['', ''] : '';
+    this.setModeAndFormat();
+
+    this.datePickerService.valueChange$.pipe(takeUntil(this.destroy$)).subscribe(() => {
+      this.updateInputValue();
+    });
   }
 
   ngOnChanges(changes: SimpleChanges): void {
@@ -294,6 +676,9 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
     if (!this.isCustomFormat) {
       this.nzFormat = inputFormats[this.nzMode as NzDateMode]!;
     }
+
+    this.inputSize = Math.max(10, this.nzFormat.length) + 2;
+    this.updateInputValue();
   }
 
   /**
@@ -302,14 +687,6 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
    */
   onOpenChange(open: boolean): void {
     this.nzOnOpenChange.emit(open);
-  }
-
-  public open(): void {
-    this.picker.showOverlay();
-  }
-
-  public close(): void {
-    this.picker.hideOverlay();
   }
 
   // ------------------------------------------------------------------------
@@ -382,7 +759,7 @@ export class NzDatePickerComponent implements OnInit, OnChanges, OnDestroy, Cont
     this.datePickerService.initialValue = newValue;
   }
 
-  onFocusChange(value: boolean): void {
+  renderClass(value: boolean): void {
     // TODO: avoid autoFocus cause change after checked error
     if (value) {
       this.renderer.addClass(this.elementRef.nativeElement, 'ant-picker-focused');
