+++ components/date-picker/date-picker.module.ts
@@ -23,7 +23,6 @@ import { InnerPopupComponent } from './inner-popup.component';
 
 import { LibPackerModule } from './lib/lib-packer.module';
 import { NzMonthPickerComponent } from './month-picker.component';
-import { NzPickerComponent } from './picker.component';
 import { NzRangePickerComponent } from './range-picker.component';
 import { NzWeekPickerComponent } from './week-picker.component';
 import { NzYearPickerComponent } from './year-picker.component';
@@ -45,7 +44,6 @@ import { NzYearPickerComponent } from './year-picker.component';
   ],
   exports: [NzDatePickerComponent, NzRangePickerComponent, NzMonthPickerComponent, NzYearPickerComponent, NzWeekPickerComponent],
   declarations: [
-    NzPickerComponent,
     NzDatePickerComponent,
     NzMonthPickerComponent,
     NzYearPickerComponent,
