+++ components/date-picker/public-api.ts
@@ -15,7 +15,6 @@ export { NzYearPickerComponent } from './year-picker.component';
 export { DatePickerService as ɵDatePickerService } from './date-picker.service';
 export { DateRangePopupComponent as ɵDateRangePopupComponent } from './date-range-popup.component';
 export { InnerPopupComponent as ɵInnerPopupComponent } from './inner-popup.component';
-export { NzPickerComponent as ɵNzPickerComponent } from './picker.component';
 export { CalendarFooterComponent as ɵCalendarFooterComponent } from './calendar-footer.component';
 
 export * from './lib/public-api';
