+++ components/date-picker/testing/util.ts
@@ -5,17 +5,12 @@
 
 import { DebugElement } from '@angular/core';
 import { By } from '@angular/platform-browser';
-import { NzPickerComponent } from '../picker.component';
 import { PREFIX_CLASS } from '../util';
 
 export const ENTER_EVENT = new KeyboardEvent('keyup', { key: 'Enter' });
 
 export function getPickerAbstract<T = HTMLElement>(debugElement: DebugElement): T {
   return debugElement.query(By.css(`.${PREFIX_CLASS}`)).nativeElement;
-}
-
-export function getPicker(debugElement: DebugElement): HTMLElement {
-  return debugElement.query(By.directive(NzPickerComponent)).nativeElement;
 }
 
 export function getPickerInput(debugElement: DebugElement): HTMLInputElement {
